refactor(phase0): PascalCase for class files (match codebase convention)
Codebase rule: a file whose primary export is a class is named after that class
(PaymentsModule.ts, TokenRegistry.ts, NametagMinter.ts); interface/function/type
files are kebab-case (oracle-provider.ts, balance-computer.ts, types.ts). Apply:
- payment-data.ts -> SpherePaymentData.ts (+ test)
- fake-engine.ts  -> FakeTokenEngine.ts (+ test)
Kept kebab: engine.ts (ITokenEngine interface), token-blob.ts (functions),
engine-contract.ts (function), sdk.ts, types.ts, index.ts. 23 tests green.
